+++ fastdeploy/model_executor/layers/quantization/__init__.py
@@ -172,7 +172,7 @@ def get_quantization_config(quantization: str) -> Type[QuantConfigBase]:
     from .weight_only import WeightOnlyConfig, WINT4Config, WINT8Config
     from .wfp8afp8 import WFP8AFP8Config
     from .wint2 import WINT2Config
-    from .compressed_tensors import CompressedTensorsConfig
+    from .compressed_tensors.compressed_tensors import CompressedTensorsConfig
 
     if envs.FD_MOE_MXFP4_BACKEND is not None:
         from .mxfp4 import MXFP4Config
