+++ fastdeploy/model_executor/layers/quantization/compressed_tensors/compressed_tensors_moe.py
@@ -1,0 +1,9 @@
+from ...linear import QuantMethodBase
+
+class CompreesedTensorsMoEMethod(QuantMethodBase):
+    
+    def create_weights():
+        pass
+    
+    def apply():
+        pass
